Moved fronted content related functions to a new service called content query...
