+++ core/Bellwether/BWCMSBundle/Classes/Content/Type/NavigationFolderType.php
@@ -88,7 +88,7 @@ class NavigationFolderType Extends ContentType
      */
     public function render($contentEntity, $options = array())
     {
-        $contentMenuItems = $this->cm()->getContentMenuItems($contentEntity);
+        $contentMenuItems = $this->cq()->getContentMenuItems($contentEntity);
         /**
          * @var \Knp\Menu\MenuItem $rootMenu
          * @var \Knp\Menu\MenuItem $menu
@@ -112,7 +112,7 @@ class NavigationFolderType Extends ContentType
             }
             if (isset($contentMeta['linkType']) && $contentMeta['linkType'] == 'content') {
                 if (isset($contentMeta['linkContent']) && ($contentMeta['linkContent'] instanceof ContentEntity)) {
-                    $contentLinkURL = $this->cm()->getPublicURL($contentMeta['linkContent']);
+                    $contentLinkURL = $this->cq()->getPublicURL($contentMeta['linkContent']);
                     $menu[$content->getId()]->setUri($contentLinkURL);
                 }
             }
