+++ core/Bellwether/BWCMSBundle/Resources/public/css/style.css
@@ -18,9 +18,16 @@ body {
     z-index: 999999999;
 }
 
+.page-header {
+    padding-bottom: 10px;
+    margin: 40px 0 0px;
+    border-bottom: 1px solid #eeeeee;
+}
+
 .page-content{
     margin-bottom: 50px;
 }
+
 
 #loadingIndicator{
     position: absolute;
@@ -43,21 +50,23 @@ body {
     margin-right: 0px;
 }
 
-div.mediaPageLeft{
+div.contentPageLeft{
+    padding-top: 10px;
     min-height: 400px;
     overflow: auto;
     background-color: #f1f0f0;
 }
 
-div.mediaPageRight {
+div.contentPageRight {
+    padding-top: 10px;
     min-height: 400px;
 }
 
-table.dataTable tr.odd {
+table.dataTable tbody tr.odd {
     background-color: #F9F9F9;
 }
 
-table.dataTable tr.even {
+table.dataTable tbody tr.even {
     background-color: #FFFFFF;
 }
